+++ binary_heap.mli
@@ -13,8 +13,19 @@
 (*                                                                        *)
 (**************************************************************************)
 
-(** Traditional implementation of priority queues
-    using a binary heap encoded in a resizable array *)
+(** Traditional implementation of priority queues using a binary heap
+    encoded in a resizable array.
+
+    When documenting complexity below, `n` refers to the number of elements
+    in the queue.
+
+    The size of the internal array is doubled when insertion requires
+    more space, and halved when less than 25% is used. As a
+    consequence, the time spent in enlarging/shrinking the array in a
+    sequence of `M` insertions (resp. deletions) is proportional to `M`,
+    and thus be considered constant time for each operation. For this
+    reason, we do not mention the possibility of a worst case complexity
+    `O(n)` in operations add/remove/pop_minimum below. *)
 
 module type Ordered = sig
   type t
@@ -39,24 +50,27 @@ module Make(X: Ordered) : sig
   (** [is_empty h] checks the emptiness of [h] *)
   val is_empty : t -> bool
 
-  (** [add x h] adds a new element [x] in heap [h]; size of [h] is doubled
-      when maximum capacity is reached; complexity $O(log(n))$ *)
+  (** [add x h] adds a new element [x] in heap [h]; complexity O(log(n)). *)
   val add : t -> X.t -> unit
 
   (** [minimum h] returns the minimum element of [h]; raises [Empty]
-      when [h] is empty; complexity $O(1)$ *)
+      when [h] is empty; complexity O(1) *)
   val minimum : t -> X.t
 
   (** [remove h] removes the minimum element of [h]; raises [Empty]
-      when [h] is empty; complexity $O(log(n))$ *)
+      when [h] is empty; complexity O(log(n)). *)
   val remove : t -> unit
 
   (** [pop_minimum h] removes the minimum element of [h] and returns it;
-      raises [Empty] when [h] is empty; complexity $O(log(n))$ *)
+      raises [Empty] when [h] is empty; complexity O(log(n)). *)
   val pop_minimum : t -> X.t
 
-  (** usual iterators and combinators; elements are presented in
-      arbitrary order *)
+  (** [remove_and_add x h] removes the minimum element of [h] and adds [x];
+      complexity O(log(n)). More efficient than calling [remove]
+      and [add]. *)
+  val remove_and_add : t -> X.t -> unit
+
+  (** usual iterators; elements are presented in arbitrary order *)
   val iter : (X.t -> unit) -> t -> unit
 
   val fold : (X.t -> 'a -> 'a) -> t -> 'a -> 'a