+++ CHANGES.md
@@ -1,7 +1,9 @@
 
+  - added `remove_and_add` (more efficient than `remove` and `add` combined)
+
 # 2.0.0
 ## changes
-  - change of API: switch a to *minimum* priority queue
+  - change of API: switch to a *minimum* priority queue
   - improved memory footprint with internal shrinking
 ## fixes
   - switch from obuild to dune and to opam 2.0